requirements.txt
google-cloud-storage
requests
PyMuPDF
google-cloud-logging
vertexai

